requirements.txt
streamlit
groq
python-dotenv==1.0.1
textblob
opencv-python
mediapipe==0.10.33
numpy
Pillow